requirements.txt
asgiref==3.6.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==40.0.2
defusedxml==0.7.1
Django==4.2
django-allauth==0.54.0
idna==3.4
numpy==1.24.3
oauthlib==3.2.2
opencv-python==4.7.0.72
Pillow==9.5.0
psycopg==3.1.8
pycparser==2.21
PyJWT==2.6.0
pyscopg2==66.0.2
python-dotenv==1.0.0
python3-openid==3.2.0
requests==2.28.2
requests-oauthlib==1.3.1
sqlparse==0.4.4
typing_extensions==4.5.0
urllib3==1.26.15
